Refactor the adapters module
