+++ src/adapters/ethereum.ts
@@ -1,5 +1,9 @@
+import { z } from "zod";
 import { Web3Provider } from "@ethersproject/providers";
-import { ethereumAccountsParser } from "src/adapters/parsers";
+
+import { StrictSchema } from "src/utils/type-safety";
+
+const ethereumAccountsParser = StrictSchema<string[]>()(z.array(z.string()));
 
 const getConnectedAccounts = (provider: Web3Provider): Promise<string[]> => {
   return provider
