+++ src/adapters/storage.ts
