+++ src/views/login/login.view.tsx
@@ -10,8 +10,8 @@ import AccountLoader from "src/views/login/components/account-loader/account-loa
 import { ReactComponent as PolygonHermezLogo } from "src/assets/polygon-hermez-logo.svg";
 import { useProvidersContext } from "src/contexts/providers.context";
 import routes from "src/routes";
-import { routerStateParser } from "src/adapters/parsers";
 import { WalletName } from "src/domain";
+import { getRouterState } from "src/adapters/browser";
 
 const Login: FC = () => {
   const classes = useLoginStyles();
@@ -30,7 +30,7 @@ const Login: FC = () => {
       setSelectedWallet(undefined);
     }
     if (account.status === "successful") {
-      const routerState = routerStateParser.safeParse(state);
+      const routerState = getRouterState(state);
 
       if (routerState.success) {
         navigate(routerState.data.redirectUrl);
